# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ayu-Trial Fabric - Core Intelligence & Governance API",
    "description": "Part B Backend & Governance Compiler API for Ayu-Trial Fabric. Powers ChangeSet impact resolution, deterministic rule evaluation, obligations lifecycle, evidence verification, and immutable audit tracking.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/v1/changesets": {
      "get": {
        "summary": "List all ChangeSets",
        "operationId": "listChangeSets",
        "responses": {
          "200": {
            "description": "List of changesets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create or propose a ChangeSet",
        "operationId": "createChangeSet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeSetRecord"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "ChangeSet created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetSingleResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          }
        }
      }
    },
    "/api/v1/changesets/{id}/impact": {
      "get": {
        "summary": "Retrieve the resolved impact graph and blast radius for a ChangeSet",
        "operationId": "getChangeSetImpact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Impact graph and metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactReportResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeSetRecord": {
        "type": "object",
        "required": ["id", "trialId", "type", "previousState", "newState", "change", "affectedEntities", "status"],
        "properties": {
          "id": { "type": "string", "example": "CS-0001" },
          "trialId": { "type": "string", "example": "ATF-001" },
          "trialName": { "type": "string" },
          "protocol": { "type": "string", "example": "v1.1" },
          "type": { "type": "string", "example": "Protocol Amendment" },
          "previousState": { "type": "string", "example": "Day 25–31" },
          "newState": { "type": "string", "example": "Day 25–35" },
          "change": { "type": "string", "example": "Visit 4 schedule: Day 25–31 → Day 25–35" },
          "affectedEntities": {
            "type": "array",
            "items": { "type": "string" }
          },
          "effectiveDate": { "type": "string" },
          "status": { "type": "string", "enum": ["SUBMITTED", "IN REVIEW", "APPROVED", "UNDER REVIEW"] },
          "created": { "type": "string" }
        }
      },
      "ImpactSummaryMetrics": {
        "type": "object",
        "required": ["sitesCount", "participantsCount", "visitsCount", "crfsCount", "edcMappingsCount", "ethicsAffected", "trainingAffected", "consentAffected"],
        "properties": {
          "sitesCount": { "type": "integer" },
          "participantsCount": { "type": "integer" },
          "visitsCount": { "type": "integer" },
          "crfsCount": { "type": "integer" },
          "edcMappingsCount": { "type": "integer" },
          "ethicsAffected": { "type": "boolean" },
          "trainingAffected": { "type": "boolean" },
          "consentAffected": { "type": "boolean" }
        }
      },
      "ImpactNode": {
        "type": "object",
        "required": ["id", "label", "entity", "category", "severity", "reason", "relationship", "relatedChangeSet"],
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" },
          "entity": { "type": "string" },
          "category": { "type": "string", "enum": ["Root", "Sites", "Participants", "Protocol", "Systems", "Governance"] },
          "severity": { "type": "string", "enum": ["HIGH", "MEDIUM", "LOW", "CRITICAL"] },
          "reason": { "type": "string" },
          "relationship": { "type": "string" },
          "relatedChangeSet": { "type": "string" },
          "parentId": { "type": "string" },
          "hasChildren": { "type": "boolean" },
          "meta": {
            "type": "object",
            "properties": {
              "count": { "type": "number" },
              "code": { "type": "string" },
              "details": { "type": "string" }
            }
          }
        }
      },
...
      "ChangeSetsResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/ChangeSetRecord" } },
          "total": { "type": "integer" },
          "timestamp": { "type": "string" }
        }
      },
      "ChangeSetSingleResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/ChangeSetRecord" },
          "timestamp": { "type": "string" }
        }
      },
      "ImpactReportResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "changeSetId": { "type": "string" },
              "trialId": { "type": "string" },
              "summary": { "$ref": "#/components/schemas/ImpactSummaryMetrics" },
              "nodes": { "type": "array", "items": { "$ref": "#/components/schemas/ImpactNode" } }
            }
          },
          "timestamp": { "type": "string" }
        }
      },
...
      "ApiErrorEnvelope": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "required": ["code", "message", "timestamp"],
            "properties": {
              "code": { "type": "string", "example": "COMPILATION_BLOCKED" },
              "message": { "type": "string", "example": "Build Failed: 3 Blockers found" },
              "details": {
                "type": "array",
                "items": { "type": "string" }
              },
              "timestamp": { "type": "string" }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequestError": {
        "description": "Invalid input or schema mismatch",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorEnvelope" }
          }
        }
      },
...
    }
  }
}